Ispravljen uslov kod brisanja stavke. Postojala je greska ako se ne obrise nijedna stavka.

+++ www/invoiceSearch.php
@@ -141,7 +141,7 @@
                 }
             }
 
-            if(!isset($_SESSION['removingItems']) || !empty($_SESSION['removingItems'])){
+            if(isset($_SESSION['removingItems']) && !empty($_SESSION['removingItems'])){
                 $sql_deleteItems = "DELETE FROM invoice_item WHERE item_id=?";
                 $stmt_deleteItems = $conn->prepare($sql_deleteItems);
                 foreach ($_SESSION['removingItems'] as $item){
